Refactoring: DiscountPercent is float now

+++ Atlas.Application/CQRS/Promos/Commands/CreatePromo/CreatePromoCommand.cs
@@ -11,7 +11,7 @@ namespace Atlas.Application.CQRS.Promos.Commands.CreatePromo
 
         public int DiscountPrice { get; set; }
 
-        public int DiscountPercent { get; set; }
+        public float DiscountPercent { get; set; }
 
         public bool ForAllGoods { get; set; }
