+++ Atlas.Presentation/Atlas.WebApi/Models/UpdatePromoDto.cs
@@ -15,7 +15,7 @@ namespace Atlas.WebApi.Models
 
         public int DiscountPrice { get; set; }
 
-        public int DiscountPercent { get; set; }
+        public float DiscountPercent { get; set; }
 
         public bool ForAllGoods { get; set; }
